+++ src/Core/Generator/VerilogModuleGenerator.cc
@@ -22,7 +22,6 @@
 
 #include "Core/Generator/VerilogModuleGenerator.h"
 #include <boost/format.hpp>
-#include <boost/foreach.hpp>
 #include <boost/algorithm/string/join.hpp>
 
 using namespace boost;
@@ -167,7 +166,7 @@ std::string VerilogModuleGenerator::generate_impl(std::string const& logic_class
 
 
     // genereate wire definitions
-    BOOST_FOREACH(net_names_table::value_type const& v, nets)
+    for (const auto& v : nets)
     {
         if (!mod->exists_module_port_name(v.second))
             wire_definitions += "  wire " + v.second + ";\n";