+++ src/Core/Primitive/QuadTreeRegionIterator.h
@@ -20,16 +20,22 @@
  */
 
 #ifndef __QUADTREEREGIONITERATOR_H__
-#define	__QUADTREEREGIONITERATOR_H__
+#define __QUADTREEREGIONITERATOR_H__
 
 // #define DEBUG_SHOW_ITER 1
 
-#include "QuadTree.h"
+#include "Core/Primitive/BoundingBox.h"
+#include "Core/Utils/Iterator.h"
+
+#include <list>
 
 namespace degate
 {
-    template <typename T>
-    class RegionIterator : public std::iterator<std::forward_iterator_tag, T>
+    template<typename T>
+    class QuadTree;
+
+    template<typename T>
+    class RegionIterator : public iterator<std::forward_iterator_tag, T>
     {
     private:
         QuadTree<T>* node;
@@ -38,7 +44,7 @@ namespace degate
         typename std::list<T>::iterator children_iter;
         typename std::list<T>::iterator children_iter_end;
 
-        std::list<QuadTree<T> *> open_list;
+        std::list<QuadTree<T>*> open_list;
 
         BoundingBox search_bb;
 
@@ -67,17 +73,17 @@ namespace degate
     /**
      * Construct an iterator end.
      */
-    template <typename T>
-    RegionIterator<T>::RegionIterator() :
-        node(nullptr), done(true)
+    template<typename T>
+    RegionIterator<T>::RegionIterator() : node(nullptr),
+                                          done(true)
     {
     }
 
-    template <typename T>
-    RegionIterator<T>::RegionIterator(QuadTree<T>* node, BoundingBox const& bbox) :
-        node(nullptr),
-        done(false),
-        search_bb(bbox)
+    template<typename T>
+    RegionIterator<T>::RegionIterator(QuadTree<T>* node, BoundingBox const& bbox)
+        : node(nullptr),
+          done(false),
+          search_bb(bbox)
     {
         assert(node != nullptr);
 
@@ -87,7 +93,7 @@ namespace degate
         skip_non_matching_children();
     }
 
-    template <typename T>
+    template<typename T>
     void RegionIterator<T>::next_node()
     {
         assert(node == nullptr);
@@ -102,7 +108,7 @@ namespace degate
             do
             {
 #ifdef DEBUG_SHOW_ITER
-    debug(TM, "get head from open list");
+                debug(TM, "get head from open list");
 #endif
                 node = open_list.front();
                 open_list.pop_front();
@@ -115,25 +121,22 @@ namespace degate
                     if ((*it).box.intersects(search_bb))
                     {
 #ifdef DEBUG_SHOW_ITER
-        debug(TM, "Put into open list: %s", (*it).get_name().c_str());
+                        debug(TM, "Put into open list: %s", (*it).get_name().c_str());
 #endif
                         open_list.push_back(&*it);
                     }
                 }
 #ifdef DEBUG_SHOW_ITER
-    debug(TM, "Current node is %s", node->get_name().c_str());
+                debug(TM, "Current node is %s", node->get_name().c_str());
 #endif
                 // reset iterator for current quadtree node
                 children_iter = node->children.begin();
                 children_iter_end = node->children.end();
 
                 // the quadtree might contain empty nodes
-            }
-            while (children_iter == children_iter_end &&
-                !open_list.empty());
+            } while (children_iter == children_iter_end && !open_list.empty());
 
-            if (children_iter == children_iter_end &&
-                open_list.empty())
+            if (children_iter == children_iter_end && open_list.empty())
             {
                 done = true;
                 node = nullptr;
@@ -141,7 +144,7 @@ namespace degate
         }
     }
 
-    template <typename T>
+    template<typename T>
     void RegionIterator<T>::check_next_node()
     {
         while (!done && children_iter == children_iter_end)
@@ -151,7 +154,7 @@ namespace degate
         }
     }
 
-    template <typename T>
+    template<typename T>
     void RegionIterator<T>::next_child()
     {
         if (!done)
@@ -163,39 +166,39 @@ namespace degate
     }
 
     // if the precond is stay on a bounding-box matching child, then postcond is stay on it, too
-    template <typename T>
+    template<typename T>
     void RegionIterator<T>::skip_non_matching_children()
     {
 #ifdef DEBUG_SHOW_ITER
-    debug(TM, "in increment() done = %d, node = %p", done ? 1 : 0, node);
+        debug(TM, "in increment() done = %d, node = %p", done ? 1 : 0, node);
 #endif
         while (!done &&
-            !search_bb.intersects(
-                get_bbox_trait_selector<is_pointer<T>::value>::get_bounding_box_for_object(*children_iter)))
+               !search_bb.intersects(
+                       get_bbox_trait_selector<is_pointer<T>::value>::get_bounding_box_for_object(*children_iter)))
         {
 #ifdef DEBUG_SHOW_ITER
-      debug(TM, "iterating over children in %s", node->get_name().c_str());
+            debug(TM, "iterating over children in %s", node->get_name().c_str());
 #endif
             next_child();
         }
 #ifdef DEBUG_SHOW_ITER
-    debug(TM, "return from increment()");
+        debug(TM, "return from increment()");
 #endif
     }
 
 
-    template <typename T>
+    template<typename T>
     RegionIterator<T>& RegionIterator<T>::operator++()
     {
 #ifdef DEBUG_SHOW_ITER
-    debug(TM, "++ called");
+        debug(TM, "++ called");
 #endif
         next_child(); // one step ahead
         skip_non_matching_children();
         return (*this);
     }
 
-    template <typename T>
+    template<typename T>
     RegionIterator<T>& RegionIterator<T>::operator=(const RegionIterator& other)
     {
         node = other.node;
@@ -206,36 +209,34 @@ namespace degate
         return (*this);
     }
 
-    template <typename T>
+    template<typename T>
     bool RegionIterator<T>::operator==(const RegionIterator& other) const
     {
         if (done == true && other.done == true)
             return true;
         else
-            return (node == other.node &&
-                children_iter == other.children_iter &&
-                open_list == other.open_list &&
-                done == other.done);
+            return (node == other.node && children_iter == other.children_iter && open_list == other.open_list &&
+                    done == other.done);
     }
 
 
-    template <typename T>
+    template<typename T>
     bool RegionIterator<T>::operator!=(const RegionIterator& other) const
     {
         return !(*this == other);
     }
 
-    template <typename T>
+    template<typename T>
     T* RegionIterator<T>::operator->() const
     {
         return &*children_iter;
     }
 
-    template <typename T>
+    template<typename T>
     T RegionIterator<T>::operator*() const
     {
         return *children_iter;
     }
-}
+} // namespace degate
 
 #endif