+++ src/Core/Project/ProjectArchiver.cc
@@ -29,7 +29,6 @@
 #include <boost/filesystem/operations.hpp>
 #include <boost/filesystem/path.hpp>
 #include <boost/format.hpp>
-#include <boost/foreach.hpp>
 
 /*
 using namespace std;