+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0-beta.2] - 202X-XX-XX
+### Added
+- New loading tiles for "attached" project mode.
+
+### Changed
+- Dark theme is now by default on Windows if desktop mode is on dark mode (#32). 
+
+### Fixed
+- Fixed crash when handling large images with new "attached" project mode (#24).
+
 ## [2.1.0-beta.1] - 2023-10-01
 ### Added
 - [Beta] New "attached" project type (optimized for huge images).